im publish 接口部分实现 //todo 需要先实现聊天记录的接口

+++ lingua_exchange/internal/types/message_types.go
@@ -148,6 +148,6 @@ type Item struct {
 type AuthOption struct {
 	TalkType          int
 	UserId            int
-	ReceiverId        int
+	ReceiverId        uint64
 	IsVerifyGroupMute bool
 }
